moved file functions into own module
