+++ cx/garbage_collector.go
@@ -72,7 +72,7 @@ func MarkAndCompact(prgrm *ast.CXProgram) {
 			offset := ptr.Offset
 			offset += fp
 
-			ptrIsPointer := IsPointer(ptr)
+			ptrIsPointer := ast.IsPointer(ptr)
 
 			// Checking if we need to mark `ptr`.
 			if ptrIsPointer {
@@ -97,7 +97,7 @@ func MarkAndCompact(prgrm *ast.CXProgram) {
 
 			// Checking if the field being accessed needs to be marked.
 			// If the root (`ptr`) is a pointer, this step is unnecessary.
-			if len(ptr.Fields) > 0 && !ptrIsPointer && IsPointer(ptr.Fields[len(ptr.Fields)-1]) {
+			if len(ptr.Fields) > 0 && !ptrIsPointer && ast.IsPointer(ptr.Fields[len(ptr.Fields)-1]) {
 				fld := ptr.Fields[len(ptr.Fields)-1]
 				MarkObjectsTree(prgrm, offset+fld.Offset, fld.Type, fld.DeclarationSpecifiers[1:])
 			}
@@ -381,7 +381,7 @@ func updatePointers(prgrm *ast.CXProgram, oldAddr, newAddr int32) {
 			// If `ptr` has fields, we need to navigate the heap and mark its fields too.
 			if glbl.CustomType != nil {
 				for _, fld := range glbl.CustomType.Fields {
-					if !IsPointer(fld) {
+					if !ast.IsPointer(fld) {
 						continue
 					}
 					offset := glbl.Offset + fld.Offset
@@ -423,7 +423,7 @@ func updatePointers(prgrm *ast.CXProgram, oldAddr, newAddr int32) {
 			offset := ptr.Offset
 			offset += fp
 
-			ptrIsPointer := IsPointer(ptr)
+			ptrIsPointer := ast.IsPointer(ptr)
 
 			// Checking if we need to mark `ptr`.
 			if ptrIsPointer {
@@ -450,7 +450,7 @@ func updatePointers(prgrm *ast.CXProgram, oldAddr, newAddr int32) {
 
 			// Checking if the field being accessed needs to be marked.
 			// If the root (`ptr`) is a pointer, this step is unnecessary.
-			if len(ptr.Fields) > 0 && !ptrIsPointer && IsPointer(ptr.Fields[len(ptr.Fields)-1]) {
+			if len(ptr.Fields) > 0 && !ptrIsPointer && ast.IsPointer(ptr.Fields[len(ptr.Fields)-1]) {
 				fld := ptr.Fields[len(ptr.Fields)-1]
 
 				// Getting the offset to the object in the heap
