+++ cx/op_und.go
@@ -214,7 +214,7 @@ func buildString(expr *ast.CXExpression, fp int) []byte {
 					res = append(res, []byte(strconv.FormatFloat(ReadF64(fp, inp), 'f', 16, 64))...)
 				}
 			case 'v':
-				res = append(res, []byte(GetPrintableValue(fp, inp))...)
+				res = append(res, []byte(ast.GetPrintableValue(fp, inp))...)
 			}
 			c++
 			specifiersCounter++
@@ -241,9 +241,9 @@ func buildString(expr *ast.CXExpression, fp int) []byte {
 			}
 
 			if c == lInps-1 {
-				extra += fmt.Sprintf("%s=%s", typ, GetPrintableValue(fp, elt))
+				extra += fmt.Sprintf("%s=%s", typ, ast.GetPrintableValue(fp, elt))
 			} else {
-				extra += fmt.Sprintf("%s=%s, ", typ, GetPrintableValue(fp, elt))
+				extra += fmt.Sprintf("%s=%s, ", typ, ast.GetPrintableValue(fp, elt))
 			}
 
 		}
