+++ cx/utilities.go
@@ -5,11 +5,7 @@ import (
 	"github.com/skycoin/cx/cx/ast"
 	"github.com/skycoin/cx/cx/constants"
 	"github.com/skycoin/cx/cx/helper"
-	"io/ioutil"
 	"os"
-	"path/filepath"
-	"strings"
-	"text/tabwriter"
 )
 
 // Debug ...
@@ -62,14 +58,14 @@ func (cxprogram *ast.CXProgram) PrintStack() {
 
 		for _, inp := range op.Inputs {
 			fmt.Println("ProgramInput")
-			fmt.Printf("\t%s : %s() : %s\n", stackValueHeader(inp.FileName, inp.FileLine), op.Name, GetPrintableValue(fp, inp))
+			fmt.Printf("\t%s : %s() : %s\n", stackValueHeader(inp.FileName, inp.FileLine), op.Name, ast.GetPrintableValue(fp, inp))
 
 			dupNames = append(dupNames, inp.Package.Name+inp.Name)
 		}
 
 		for _, out := range op.Outputs {
 			fmt.Println("ProgramOutput")
-			fmt.Printf("\t%s : %s() : %s\n", stackValueHeader(out.FileName, out.FileLine), op.Name, GetPrintableValue(fp, out))
+			fmt.Printf("\t%s : %s() : %s\n", stackValueHeader(out.FileName, out.FileLine), op.Name, ast.GetPrintableValue(fp, out))
 
 			dupNames = append(dupNames, out.Package.Name+out.Name)
 		}
@@ -95,7 +91,7 @@ func (cxprogram *ast.CXProgram) PrintStack() {
 				// fmt.Println("\t", inp.Name, "\t", ":", "\t", GetPrintableValue(fp, inp))
 				// exprs += fmt.Sprintln("\t", stackValueHeader(inp.FileName, inp.FileLine), "\t", ":", "\t", GetPrintableValue(fp, inp))
 
-				exprs += fmt.Sprintf("\t%s : %s() : %s\n", stackValueHeader(inp.FileName, inp.FileLine), ExprOpName(expr), GetPrintableValue(fp, inp))
+				exprs += fmt.Sprintf("\t%s : %s() : %s\n", stackValueHeader(inp.FileName, inp.FileLine), ExprOpName(expr), ast.GetPrintableValue(fp, inp))
 
 				dupNames = append(dupNames, inp.Package.Name+inp.Name)
 			}
@@ -118,7 +114,7 @@ func (cxprogram *ast.CXProgram) PrintStack() {
 				// fmt.Println("\t", out.Name, "\t", ":", "\t", GetPrintableValue(fp, out))
 				// exprs += fmt.Sprintln("\t", stackValueHeader(out.FileName, out.FileLine), ":", GetPrintableValue(fp, out))
 
-				exprs += fmt.Sprintf("\t%s : %s() : %s\n", stackValueHeader(out.FileName, out.FileLine), ExprOpName(expr), GetPrintableValue(fp, out))
+				exprs += fmt.Sprintf("\t%s : %s() : %s\n", stackValueHeader(out.FileName, out.FileLine), ExprOpName(expr), ast.GetPrintableValue(fp, out))
 
 				dupNames = append(dupNames, out.Package.Name+out.Name)
 			}
@@ -566,393 +562,3 @@ func SliceRemove(fp int, out *ast.CXArgument, inp *ast.CXArgument, index int32, 
 	return int(outputSliceOffset)
 }
 
-// WriteToSlice is used to create slices in the backend, i.e. not by calling `append`
-// in a CX program, but rather by the CX code itself. This function is used by
-// affordances, serialization and to store OS input arguments.
-func WriteToSlice(off int, inp []byte) int {
-	// TODO: Check all these parses from/to int32/int.
-	var inputSliceLen int32
-	if off != 0 {
-		inputSliceLen = GetSliceLen(int32(off))
-	}
-
-	inpLen := len(inp)
-	// We first check if a resize is needed. If a resize occurred
-	// the address of the new slice will be stored in `newOff` and will
-	// be different to `off`.
-	newOff := SliceResizeEx(int32(off), inputSliceLen+1, inpLen)
-
-	// Copy the data from the old slice at `off` to `newOff`.
-	SliceCopyEx(int32(newOff), int32(off), inputSliceLen+1, inpLen)
-
-	// Write the new slice element `inp` to the slice located at `newOff`.
-	SliceAppendWrite(int32(newOff), inp, inputSliceLen)
-	return newOff
-
-}
-
-func getNonCollectionValue(fp int, arg, elt *ast.CXArgument, typ string) string {
-	switch typ {
-	case "bool":
-		return fmt.Sprintf("%v", ReadBool(fp, elt))
-	case "str":
-		return fmt.Sprintf("%v", ReadStr(fp, elt))
-	case "i8":
-		return fmt.Sprintf("%v", ReadI8(fp, elt))
-	case "i16":
-		return fmt.Sprintf("%v", ReadI16(fp, elt))
-	case "i32":
-		return fmt.Sprintf("%v", ReadI32(fp, elt))
-	case "i64":
-		return fmt.Sprintf("%v", ReadI64(fp, elt))
-	case "ui8":
-		return fmt.Sprintf("%v", ReadUI8(fp, elt))
-	case "ui16":
-		return fmt.Sprintf("%v", ReadUI16(fp, elt))
-	case "ui32":
-		return fmt.Sprintf("%v", ReadUI32(fp, elt))
-	case "ui64":
-		return fmt.Sprintf("%v", ReadUI64(fp, elt))
-	case "f32":
-		return fmt.Sprintf("%v", ReadF32(fp, elt))
-	case "f64":
-		return fmt.Sprintf("%v", ReadF64(fp, elt))
-	default:
-		// then it's a struct
-		var val string
-		val = "{"
-		// for _, fld := range elt.CustomType.Fields {
-		lFlds := len(elt.CustomType.Fields)
-		off := 0
-		for c := 0; c < lFlds; c++ {
-			fld := elt.CustomType.Fields[c]
-			if c == lFlds-1 {
-				val += fmt.Sprintf("%s: %s", fld.Name, GetPrintableValue(fp+arg.Offset+off, fld))
-			} else {
-				val += fmt.Sprintf("%s: %s, ", fld.Name, GetPrintableValue(fp+arg.Offset+off, fld))
-			}
-			off += fld.TotalSize
-		}
-		val += "}"
-		return val
-	}
-}
-
-// GetPrintableValue ...
-func GetPrintableValue(fp int, arg *ast.CXArgument) string {
-	var typ string
-	elt := GetAssignmentElement(arg)
-	if elt.CustomType != nil {
-		// then it's custom type
-		typ = elt.CustomType.Name
-	} else {
-		// then it's native type
-		typ = constants.TypeNames[elt.Type]
-	}
-
-	if len(elt.Lengths) > 0 {
-		var val string
-		if len(elt.Lengths) == 1 {
-			val = "["
-			for c := 0; c < elt.Lengths[0]; c++ {
-				if c == elt.Lengths[0]-1 {
-					val += getNonCollectionValue(fp+c*elt.Size, arg, elt, typ)
-				} else {
-					val += getNonCollectionValue(fp+c*elt.Size, arg, elt, typ) + ", "
-				}
-
-			}
-			val += "]"
-		} else {
-			// 5, 4, 1
-			val = ""
-
-			finalSize := 1
-			for _, l := range elt.Lengths {
-				finalSize *= l
-			}
-
-			lens := make([]int, len(elt.Lengths))
-			copy(lens, elt.Lengths)
-
-			for c := 0; c < len(lens); c++ {
-				for i := 0; i < len(lens[c+1:]); i++ {
-					lens[c] *= lens[c+i]
-				}
-			}
-
-			for range lens {
-				val += "["
-			}
-
-			// adding first element because of formatting reasons
-			val += getNonCollectionValue(fp, arg, elt, typ)
-			for c := 1; c < finalSize; c++ {
-				closeCount := 0
-				for _, l := range lens {
-					if c%l == 0 && c != 0 {
-						// val += "] ["
-						closeCount++
-					}
-				}
-
-				if closeCount > 0 {
-					for c := 0; c < closeCount; c++ {
-						val += "]"
-					}
-					val += " "
-					for c := 0; c < closeCount; c++ {
-						val += "["
-					}
-
-					val += getNonCollectionValue(fp+c*elt.Size, arg, elt, typ)
-				} else {
-					val += " " + getNonCollectionValue(fp+c*elt.Size, arg, elt, typ)
-				}
-			}
-			for range lens {
-				val += "]"
-			}
-		}
-
-		return val
-	}
-
-	return getNonCollectionValue(fp, arg, elt, typ)
-}
-
-// DebugHeap prints the symbols that are acting as pointers in a CX program at certain point during the execution of the program along with the addresses they are pointing. Additionally, a list of the objects in the heap is printed, which shows their address in the heap, if they are marked as alive or as dead by the garbage collector, the address where they used to live after a garbage collector call, the full size of the object, the object itself as a slice of bytes and the pointers that are pointing to that object.
-func DebugHeap() {
-	// symsToAddrs will hold a list of symbols that are pointing to an address.
-	symsToAddrs := make(map[int32][]string)
-
-	// Processing global variables. Adding the address they are pointing to.
-	for _, pkg := range ast.PROGRAM.Packages {
-		for _, glbl := range pkg.Globals {
-			if glbl.IsPointer || glbl.IsSlice {
-				heapOffset := helper.Deserialize_i32(ast.PROGRAM.Memory[glbl.Offset : glbl.Offset+constants.TYPE_POINTER_SIZE])
-
-				symsToAddrs[heapOffset] = append(symsToAddrs[heapOffset], glbl.Name)
-			}
-		}
-	}
-
-	// Processing local variables in every active function call in the `CallStack`.
-	// Adding the address they are pointing to.
-	var fp int
-	for c := 0; c <= ast.PROGRAM.CallCounter; c++ {
-		op := ast.PROGRAM.CallStack[c].Operator
-
-		// TODO: Some standard library functions "manually" add a function
-		// call (callbacks) to `PRGRM.CallStack`. These functions do not have an
-		// operator associated to them. This can be considered as a bug or as an
-		// undesirable mechanic.
-		// [2019-06-24 Mon 22:39] Actually, if the GC is triggered in the middle
-		// of a callback, things will certainly break.
-		if op == nil {
-			continue
-		}
-
-		for _, ptr := range op.ListOfPointers {
-			offset := ptr.Offset
-			symName := ptr.Name
-			if len(ptr.Fields) > 0 {
-				fld := ptr.Fields[len(ptr.Fields)-1]
-				offset += fld.Offset
-				symName += "." + fld.Name
-			}
-
-			if ptr.Offset < ast.PROGRAM.StackSize {
-				offset += fp
-			}
-
-			heapOffset := helper.Deserialize_i32(ast.PROGRAM.Memory[offset : offset+constants.TYPE_POINTER_SIZE])
-
-			symsToAddrs[heapOffset] = append(symsToAddrs[heapOffset], symName)
-		}
-
-		fp += op.Size
-	}
-
-	// Printing all the details.
-	w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, '.', 0)
-
-	for off, symNames := range symsToAddrs {
-		var addrB [4]byte
-		WriteMemI32(addrB[:], 0, off)
-		fmt.Fprintln(w, "Addr:\t", addrB, "\tPtr:\t", symNames)
-	}
-
-	// Just a newline.
-	fmt.Fprintln(w)
-	w.Flush()
-
-	w = tabwriter.NewWriter(os.Stdout, 0, 0, 2, '.', 0)
-
-	for c := ast.PROGRAM.HeapStartsAt + constants.NULL_HEAP_ADDRESS_OFFSET; c < ast.PROGRAM.HeapStartsAt+ast.PROGRAM.HeapPointer; {
-		objSize := helper.Deserialize_i32(ast.PROGRAM.Memory[c+constants.MARK_SIZE+constants.FORWARDING_ADDRESS_SIZE : c+constants.MARK_SIZE+constants.FORWARDING_ADDRESS_SIZE+constants.OBJECT_SIZE])
-
-		// Setting a limit size for the object to be printed if the object is too large.
-		// We don't want to print obscenely large objects to standard output.
-		printObjSize := objSize
-		if objSize > 50 {
-			printObjSize = 50
-		}
-
-		var addrB [4]byte
-		WriteMemI32(addrB[:], 0, int32(c))
-
-		fmt.Fprintln(w, "Addr:\t", addrB, "\tMark:\t", ast.PROGRAM.Memory[c:c+constants.MARK_SIZE], "\tFwd:\t", ast.PROGRAM.Memory[c+constants.MARK_SIZE:c+constants.MARK_SIZE+constants.FORWARDING_ADDRESS_SIZE], "\tSize:\t", objSize, "\tObj:\t", ast.PROGRAM.Memory[c+constants.OBJECT_HEADER_SIZE:c+int(printObjSize)], "\tPtrs:", symsToAddrs[int32(c)])
-
-		c += int(objSize)
-	}
-
-	// Just a newline.
-	fmt.Fprintln(w)
-	w.Flush()
-}
-
-// filePathWalkDir scans all the files in a directory. It will automatically
-// scan each sub-directories in the directory. Code obtained from manigandand's
-// post in https://stackoverflow.com/questions/14668850/list-directory-in-go.
-func filePathWalkDir(root string) ([]string, error) {
-	var files []string
-	err := filepath.Walk(root, func(path string, info os.FileInfo, err error) error {
-		if !info.IsDir() {
-			files = append(files, path)
-		}
-		return err
-	})
-	return files, err
-}
-
-// ioReadDir reads the directory named by dirname and returns a list of
-// directory entries sorted by filename. Code obtained from manigandand's
-// post in https://stackoverflow.com/questions/14668850/list-directory-in-go.
-func ioReadDir(root string) ([]string, error) {
-	var files []string
-	fileInfo, err := ioutil.ReadDir(root)
-	if err != nil {
-		return files, err
-	}
-
-	for _, file := range fileInfo {
-		files = append(files, fmt.Sprintf("%s/%s", root, file.Name()))
-	}
-	return files, nil
-}
-
-// ParseArgsForCX parses the arguments and returns:
-//  - []arguments
-//  - []file pointers	open files
-//  - []sting		filenames
-func ParseArgsForCX(args []string, alsoSubdirs bool) (cxArgs []string, sourceCode []*os.File, fileNames []string) {
-	skip := false // flag for skipping arg
-
-	for _, arg := range args {
-
-		// skip arg if skip flag is specified
-		if skip {
-			skip = false
-			continue
-		}
-
-		// cli flags are either "--key=value" or "-key value"
-		// we have to skip both cases
-		if len(arg) > 1 && arg[0] == '-' {
-			if !strings.Contains(arg, "=") {
-				skip = true
-			}
-			continue
-		}
-
-		// cli cx flags are prefixed with "++"
-		if len(arg) > 2 && arg[:2] == "++" {
-			cxArgs = append(cxArgs, arg)
-			continue
-		}
-
-		fi, err := CXStatFile(arg)
-		if err != nil {
-			println(fmt.Sprintf("%s: source file or library not found", arg))
-			os.Exit(constants.CX_COMPILATION_ERROR)
-		}
-
-		switch mode := fi.Mode(); {
-		case mode.IsDir():
-			var fileList []string
-			var err error
-
-			// Checking if we want to check all subdirectories.
-			if alsoSubdirs {
-				fileList, err = filePathWalkDir(arg)
-			} else {
-				fileList, err = ioReadDir(arg)
-				// fileList, err = filePathWalkDir(arg)
-			}
-
-			if err != nil {
-				panic(err)
-			}
-
-			for _, path := range fileList {
-				file, err := CXOpenFile(path)
-
-				if err != nil {
-					println(fmt.Sprintf("%s: source file or library not found", arg))
-					os.Exit(constants.CX_COMPILATION_ERROR)
-				}
-
-				fiName := file.Name()
-				fiNameLen := len(fiName)
-
-				if fiNameLen > 2 && fiName[fiNameLen-3:] == ".cx" {
-					// only loading .cx files
-					sourceCode = append(sourceCode, file)
-					fileNames = append(fileNames, fiName)
-				}
-			}
-		case mode.IsRegular():
-			file, err := CXOpenFile(arg)
-
-			if err != nil {
-				panic(err)
-			}
-
-			fileNames = append(fileNames, file.Name())
-			sourceCode = append(sourceCode, file)
-		}
-	}
-
-	return cxArgs, sourceCode, fileNames
-}
-
-// IsPointer checks if `sym` is a candidate for the garbage collector to check.
-// For example, if `sym` is a slice, the garbage collector will need to check
-// if the slice on the heap needs to be relocated.
-func IsPointer(sym *ast.CXArgument) bool {
-	// There's no need to add global variables in `fn.ListOfPointers` as we can access them easily through `CXPackage.Globals`
-	// TODO: We could still pre-compute a list of candidates for globals.
-	if sym.Offset >= ast.PROGRAM.StackSize && sym.Name != "" {
-		return false
-	}
-	// NOTE: Strings are considered as `IsPointer`s by the runtime.
-	// if (sym.IsPointer || sym.IsSlice) && sym.Name != "" {
-	// 	return true
-	// }
-	if (sym.IsPointer || sym.IsSlice) && sym.Name != "" && len(sym.Fields) == 0 {
-		return true
-	}
-	if sym.Type == constants.TYPE_STR && sym.Name != "" && len(sym.Fields) == 0 {
-		return true
-	}
-	// if (sym.Type == TYPE_STR && sym.Name != "") {
-	// 	return true
-	// }
-	// If `sym` is a structure instance, we need to check if the last field
-	// being access is a pointer candidate
-	// if len(sym.Fields) > 0 {
-	// 	return isPointer(sym.Fields[len(sym.Fields)-1])
-	// }
-	return false
-}
