+++ cxgo/cxparser/cxparser.go
@@ -241,7 +241,7 @@ func lexerStep0(srcStrs, srcNames []string) int {
 					// Checking if `pkgName` already exists and if it's not a standard library package.
 					if _, err := cxgo0.PRGRM0.GetPackage(pkgName); err != nil && !cxcore.IsCorePackage(pkgName) {
 						// _, sourceCode, srcNames := ParseArgsForCX([]string{fmt.Sprintf("%s%s", SRCPATH, pkgName)}, false)
-						_, sourceCode, fileNames := cxcore.ParseArgsForCX([]string{filepath.Join(globals2.SRCPATH, pkgName)}, false)
+						_, sourceCode, fileNames := ast.ParseArgsForCX([]string{filepath.Join(globals2.SRCPATH, pkgName)}, false)
 						ParseSourceCode(sourceCode, fileNames)
 					}
 				}
