+++ cxgo/util/profiling/profiling.go
@@ -2,12 +2,11 @@ package profiling
 
 import (
 	"fmt"
+	"github.com/skycoin/cx/cx/util/file"
 	"os"
 	"runtime"
 	"runtime/pprof"
 	"time"
-
-	"github.com/skycoin/cx/cx"
 )
 
 var (
@@ -35,7 +34,7 @@ func StopProfile(name string) {
 
 func StartCPUProfile(name string) *os.File {
 	if DebugProfile {
-		f, err := cxcore.CXCreateFile(fmt.Sprintf("%s_%s_cpu.pprof", os.Args[0], name))
+		f, err := file.CXCreateFile(fmt.Sprintf("%s_%s_cpu.pprof", os.Args[0], name))
 		if err != nil {
 			fmt.Println("Failed to create CPU profile: ", err)
 		}
@@ -61,7 +60,7 @@ func StopCPUProfile(f *os.File) {
 
 func DumpMEMProfile(name string) {
 	if DebugProfile {
-		f, err := cxcore.CXCreateFile(fmt.Sprintf("%s_%s_mem.pprof", os.Args[0], name))
+		f, err := file.CXCreateFile(fmt.Sprintf("%s_%s_mem.pprof", os.Args[0], name))
 		if err != nil {
 			fmt.Println("Failed to create MEM profile: ", err)
 		}
