+++ cxos/op_profile.go
@@ -4,6 +4,7 @@ package cxos
 
 import (
 	"fmt"
+	"github.com/skycoin/cx/cx/util/file"
 	"os"
 	"runtime"
 	"runtime/pprof"
@@ -14,7 +15,7 @@ import (
 var openProfiles map[string]*os.File = make(map[string]*os.File, 0)
 
 func startCPUProfile(name string, rate int) *os.File {
-	f, err := cxcore.CXCreateFile(fmt.Sprintf("%s_%s_cpu.pprof", os.Args[0], name))
+	f, err := file.CXCreateFile(fmt.Sprintf("%s_%s_cpu.pprof", os.Args[0], name))
 	if err != nil {
 		fmt.Println("Failed to create CPU profile: ", err)
 	}
